# sql/parser: trim sqlSymType and Scanner structs