--- a/pkg/sql/parser/sql.y
+++ b/pkg/sql/parser/sql.y
@@ -525,7 +525,6 @@ func (u *sqlSymUnion) scrubOption() tree.ScrubOption {
 %union {
   id             int
   pos            int
-  empty          struct{}
   str            string
   union          sqlSymUnion
 }