@knz 2017-12-18
LGTM on principle but CI says you broke contextual help.

---

Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20834#-:-L0faBnY7fpNiXDcExxD:b-rdojjz)*
<!-- Sent from Reviewable.io -->